refactor: rename helpers dir to utils
