+++ package.json
@@ -139,6 +139,7 @@
     "slugify": "^1.6.4",
     "swingset": "^0.11.0",
     "unified": "^9.2.0",
+    "unist-util-is": "^5.1.1",
     "unist-util-visit": "^4.1.0",
     "validator": "^13.7.0"
   },
