+++ src/views/tutorial-view/utils/serialize-content.ts
